backend/requirements.txt
fastapi
uvicorn
requests
readability-lxml
beautifulsoup4
lxml
python-docx
python-multipart
sse-starlette
